+++ iti-react-core/src/CancellablePromise/CancellablePromise.ts
@@ -32,9 +32,11 @@ export class CancellablePromise<T> {
         // The returned promise should resolve even after its canceled.
         // The idea is that the promise is resolved instantaneously, so by the time
         // the promise is canceled, it has already resolved.
-        return new CancellablePromise(Promise.resolve(value), () => {
-            /* no-op */
-        })
+        return new CancellablePromise(Promise.resolve(value), noop)
+    }
+
+    static reject<T>(reason?: unknown): CancellablePromise<T> {
+        return new CancellablePromise(Promise.reject(reason), noop)
     }
 
     static all<T1>(promises: [CancellablePromise<T1>]): CancellablePromise<[T1]>
@@ -154,40 +156,4 @@ export class CancellablePromise<T> {
             rejectFn()
         })
     }
-}
-
-export const PSEUDO_PROMISE_CANCELED = 'PSEUDO_PROMISE_CANCELED'
-
-export function pseudoCancellable<T>(promise: PromiseLike<T>): CancellablePromise<T> {
-    let canceled = false
-
-    const wrappedPromise = promise.then(result => {
-        if (canceled) throw new Error(PSEUDO_PROMISE_CANCELED)
-        return result
-    })
-
-    return new CancellablePromise(wrappedPromise, () => {
-        canceled = true
-    })
-}
-
-export type CaptureCancellablePromise = <T>(
-    promise: CancellablePromise<T>
-) => CancellablePromise<T>
-
-export function buildCancellablePromise<T>(
-    innerFunc: (capture: CaptureCancellablePromise) => PromiseLike<T>
-): CancellablePromise<T> {
-    const capturedPromises: CancellablePromise<unknown>[] = []
-
-    const capture: CaptureCancellablePromise = promise => {
-        capturedPromises.push(promise)
-        return promise
-    }
-
-    function cancel(): void {
-        for (const promise of capturedPromises) promise.cancel()
-    }
-
-    return new CancellablePromise(innerFunc(capture), cancel)
 }