+++ iti-react-core/src/__tests__/CancellablePromise/CancellablePromise.test.ts
@@ -1,18 +1,13 @@
 ﻿import 'jest'
 import { range } from 'lodash'
 import { performance } from 'perf_hooks'
-import {
-    CancellablePromise,
-    pseudoCancellable,
-    PSEUDO_PROMISE_CANCELED,
-    buildCancellablePromise
-} from '../CancellablePromise'
+import { CancellablePromise } from '../../CancellablePromise'
 
 interface Options {
     resolve?: boolean
 }
 
-function getPromise<T>(
+export function getPromise<T>(
     returnValue: T,
     duration: number,
     options: Options = { resolve: true }
@@ -190,6 +185,35 @@ describe('CancellablePromise.resolve', () => {
     })
 })
 
+describe('CancellablePromise.reject', () => {
+    test('reject', async () => {
+        try {
+            await CancellablePromise.reject<number>(7)
+            fail('Promise resolved.')
+        } catch (e) {
+            expect(e).toBe(7)
+        }
+    })
+
+    test('reject<void>', async () => {
+        try {
+            const pVoid: CancellablePromise<void> = CancellablePromise.reject()
+            await pVoid
+            fail('Promise resolved.')
+        } catch (e) {
+            expect(e).toBeUndefined()
+        }
+
+        try {
+            const pVoid2: CancellablePromise<void> = CancellablePromise.reject<void>()
+            await pVoid2
+            fail('Promise resolved.')
+        } catch (e) {
+            expect(e).toBeUndefined()
+        }
+    })
+})
+
 describe('CancellablePromise.delay', () => {
     it('delays', async () => {
         const start = performance.now()
@@ -202,83 +226,5 @@ describe('CancellablePromise.delay', () => {
         const promise = CancellablePromise.delay(200)
         promise.cancel()
         await assertRejects(promise)
-    })
-})
-
-describe('pseudoCancellable', () => {
-    test('success', async () => {
-        try {
-            const number = await pseudoCancellable(Promise.resolve(1))
-            expect(number).toBe(1)
-        } catch (e) {
-            fail('Promise rejected.')
-        }
-    })
-
-    test('canceled', async () => {
-        try {
-            const promise = pseudoCancellable(
-                new Promise(resolve => setTimeout(resolve, 1000))
-            )
-
-            await getPromise('', 100)
-            promise.cancel()
-            await promise
-
-            fail('Promise resolved when it should not have.')
-        } catch (e) {
-            expect(e.message).toBe(PSEUDO_PROMISE_CANCELED)
-        }
-    })
-})
-
-describe('buildCancellablePromise', () => {
-    test('cancellation works', async () => {
-        const promise1Duration = 100
-
-        const overallPromise = buildCancellablePromise(async capture => {
-            await capture(getPromise('1', promise1Duration))
-
-            try {
-                await capture(getPromise('2', 3 * promise1Duration))
-                fail('Promise 2 resolved when it should have been canceled.')
-            } catch {
-                // do nothing
-            }
-        })
-
-        // Wait until promise1 resolves
-        await getPromise('', 2 * promise1Duration)
-
-        // This should cause promise2 to be canceled
-        overallPromise.cancel()
-    })
-
-    test('simultaneous promises captured', async () => {
-        const duration = 100
-
-        const overallPromise = buildCancellablePromise(async capture => {
-            const promise1 = capture(getPromise('1', duration))
-            const promise2 = capture(getPromise('2', duration))
-
-            try {
-                await promise1
-                fail('promise1 resolved')
-            } catch {
-                // do nothing
-            }
-
-            try {
-                await promise2
-                fail('promise2 resolved')
-            } catch {
-                // do nothing
-            }
-        })
-
-        overallPromise.cancel()
-
-        // Wait to see if either promise resolves
-        await getPromise('', duration * 2)
     })
 })