+++ iti-react-core/src/__tests__/__helpers__/TestItiReactCoreContextData.ts
@@ -1,0 +1,16 @@
+import {
+    ItiReactCoreContextData,
+    defaultItiReactCoreContextData
+} from '../../ItiReactCoreContext'
+
+export const testItiReactCoreContextData: ItiReactCoreContextData = {
+    onError: e => {
+        throw e
+    },
+    useAutoRefreshQuery: {
+        ...defaultItiReactCoreContextData.useAutoRefreshQuery,
+        isConnectionError: e => {
+            throw e
+        }
+    }
+}